refactor(client): extract useTaskEditForm from the task edit route
Moves the task edit page's data fetching, form construction, submit and
delete wiring into a route-private useTaskEditForm hook (now using the
queryKeys factory instead of inline key arrays), leaving the route a thin
JSX shell. Clears the file's import/max-dependencies warning (12 -> 5
value imports).
